Created color picker component in ads
Added theme changes in color picker component
